Delete outdated tsne.go and update Dockerfile for new wasm build
Removed tsne.go since its functionality is now incorporated within memory.go. Updated Dockerfile to correctly build the new wasm file and change working directories accordingly. This commit streamlines the build process and centralizes memory management code.
